+++ Apocalypse.Any.Client/GameInstance/ClientWorldGame.cs
@@ -33,7 +33,7 @@ namespace Apocalypse.Any.Client.GameInstance
         {
             SerializationAdapter = serializationAdapter ?? throw new ArgumentNullException(nameof(serializationAdapter));
         }
-        private NetOutgoingMessage CreateMessage<T>(int commandName, T instanceToSend)
+        private NetOutgoingMessage CreateMessage<T>(byte commandName, T instanceToSend)
         {
             return Client.CreateMessage(
